feat: bmi_calculator: Add input_page module
