+++ lib/input_page.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class InputPage extends StatefulWidget {
+  const InputPage({super.key});
+
+  @override
+  State<InputPage> createState() => _InputPageState();
+}
+
+class _InputPageState extends State<InputPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Center(
+          child: Text('BMI Calculator'),
+        ),
+      ),
+      body: const SafeArea(
+        child: Center(
+          child: Text('Body here'),
+        ),
+      ),
+    );
+  }
+}
