+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'input_page.dart';
 
 void main() {
   runApp(const BMICalculator());
@@ -17,31 +18,6 @@ class BMICalculator extends StatelessWidget {
         scaffoldBackgroundColor: const Color(0xff0a0e21),
       ),
       home: const InputPage(),
-    );
-  }
-}
-
-class InputPage extends StatefulWidget {
-  const InputPage({super.key});
-
-  @override
-  State<InputPage> createState() => _InputPageState();
-}
-
-class _InputPageState extends State<InputPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Center(
-          child: Text('BMI Calculator'),
-        ),
-      ),
-      body: const SafeArea(
-        child: Center(
-          child: Text('Body here'),
-        ),
-      ),
     );
   }
 }
